move ChargeCurrentDistributor to SolarChargers namespace
